up and running test

+++ check_web/check_web/pipelines.py
@@ -8,6 +8,7 @@
 import json
 import codecs
 import sys
+import os
 
 reload(sys)
 sys.setdefaultencoding('utf8')
@@ -15,7 +16,11 @@ sys.setdefaultencoding('utf8')
 
 class CheckWebPipeline(object):
     def __init__(self):
-        self.file = codecs.open('test.json', mode='wb', encoding='utf-8')
+        if(os.path.isfile('test.json')):
+            self.file = codecs.open('test1.json', mode='wb', encoding='utf-8')
+        else:
+            self.file = codecs.open('test.json', mode='wb', encoding='utf-8')
+        # self.file = codecs.open('test.json', mode='wb', encoding='utf-8')
 
     def process_item(self, item, spider):
         line = 'the new movie list:' + '\n\n\n'
